+++ src/RefDocGen/Intermed/ClassIntermed.cs
@@ -1,0 +1,5 @@
+namespace RefDocGen.Intermed;
+
+public record ClassIntermed(string Name, AccessModifier AccessibilityModifier, FieldIntermed[] Fields, PropertyIntermed[] Properties, MethodIntermed[] Methods)
+{
+}
