+++ src/RefDocGen/Intermed/MethodParameterIntermed.cs
@@ -1,0 +1,22 @@
+using System.Reflection;
+
+namespace RefDocGen.Intermed;
+
+public record MethodParameterIntermed(ParameterInfo ParameterInfo)
+{
+    public string Name => ParameterInfo.Name;
+
+    public string Type => ParameterInfo.ParameterType.Name;
+
+    public bool IsParamsArray => ParameterInfo.GetCustomAttribute(typeof(ParamArrayAttribute)) != null;
+
+    public bool IsOptional => ParameterInfo.IsOptional;
+
+    public bool IsInput => ParameterInfo.IsIn;
+
+    public bool IsOutput => ParameterInfo.IsOut;
+
+    public bool IsPassedByReference => ParameterInfo.ParameterType.IsByRef;
+
+    public bool HasRefKeyword => IsPassedByReference && !IsInput && !IsOutput;
+}
